refactored names for classes/paths
